chore(appium): Move mixin BiDi-related methods into a separate module
