+++ packages/base-driver/lib/protocol/errors.js
@@ -60,16 +60,14 @@ export class ProtocolError extends BaseError {
    * Get the Bidi protocol version of an error
    * @param {string|number} id - the id used in the request for which this error forms the response
    * @see https://w3c.github.io/webdriver-bidi/#protocol-definition
-   * @returns The object conforming to the shape of a BiDi error response
+   * @returns {import('@appium/types').ErrorBiDiCommandResponse} The object conforming to the shape of a BiDi error response
    */
   bidiErrObject(id) {
     // if we don't have an id, the client didn't send one, so we have nothing to send back.
-    // send back an empty string rather than making something up
-    if (_.isNil(id)) {
-      id = '';
-    }
+    // send back zero rather than making something up
+    const intId = /** @type {number} */ (_.isInteger(id) ? id : (parseInt(`${id}`, 10) || 0));
     return {
-      id,
+      id: intId,
       type: 'error',
       error: this.error,
       stacktrace: this.stacktrace,
