+++ packages/types/lib/command.ts
@@ -155,3 +155,20 @@ export interface BidiMethodMap {
 export interface BidiModuleMap {
   [k: string]: BidiMethodMap;
 }
+
+// https://w3c.github.io/webdriver-bidi/#protocol-definition
+export interface GenericBiDiCommandResponse {
+  id: number;
+  type: 'success' | 'error';
+  [key: string]: any;
+}
+
+export interface SuccessBiDiCommandResponse extends GenericBiDiCommandResponse {
+  result: any;
+}
+
+export interface ErrorBiDiCommandResponse extends GenericBiDiCommandResponse {
+  error: string;
+  message: string;
+  stacktrace?: string;
+}
